moved functionality of (I)AuthorizedLinkAggregator to (I)MenuGenerator
