--- src/Sienar.Utils/Infrastructure/AuthorizedLinkAggregator.cs
@@ -1,77 +1,0 @@
-﻿#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
-
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Sienar.Infrastructure;
-
-/// <exclude />
-public class AuthorizedLinkAggregator<TLink> : IAuthorizedLinkAggregator<TLink>
-	where TLink : MenuLink
-{
-	private readonly IUserAccessor _userAccessor;
-	private readonly IDictionaryProvider<LinkDictionary<TLink>> _provider;
-
-	public AuthorizedLinkAggregator(
-		IUserAccessor userAccessor,
-		IDictionaryProvider<LinkDictionary<TLink>> provider)
-	{
-		_userAccessor = userAccessor;
-		_provider = provider;
-	}
-
-	/// <inheritdoc />
-	public Task<List<TLink>> Create(string name)
-	{
-		var linkDictionary = _provider.Access(name);
-		return ProcessNavLinks(linkDictionary.AggregatePrioritized());
-	}
-
-	protected async Task<List<TLink>> ProcessNavLinks(List<TLink> navLinks)
-	{
-		var includedLinks = new List<TLink>();
-
-		foreach (var link in navLinks)
-		{
-			if (!await UserIsAuthorized(link))
-			{
-				continue;
-			}
-
-			await PerformAdditionalProcessing(link);
-
-			includedLinks.Add(link);
-		}
-
-		return includedLinks;
-	}
-
-	protected virtual Task PerformAdditionalProcessing(TLink link) => Task.CompletedTask;
-
-	private async Task<bool> UserIsAuthorized(TLink menuLink)
-	{
-		if (menuLink.RequireLoggedIn && !await _userAccessor.IsSignedIn()) return false;
-		if (menuLink.RequireLoggedOut && await _userAccessor.IsSignedIn()) return false;
-		if (menuLink.Roles is null) return true;
-
-		foreach (var role in menuLink.Roles)
-		{
-			if (await _userAccessor.UserInRole(role))
-			{
-				if (menuLink.AllRolesRequired)
-				{
-					continue;
-				}
-
-				return true;
-			}
-
-			if (menuLink.AllRolesRequired)
-			{
-				return false;
-			}
-		}
-
-		return menuLink.AllRolesRequired;
-	}
-}
